pdf_parser/modules/pdf_extractor.py
import os
import io
import PyPDF2
from pdfminer.high_level import extract_text, extract_pages
from pdfminer.layout import LTImage, LTFigure
import binascii
import logging

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class PDFExtractor:
    """Class for extracting data from PDF documents"""

    def __init__(self, pdf_path):
        """
        Initialize with path to PDF file
        
        Args:
            pdf_path (str): Path to the PDF file to analyze
        """
        self.pdf_path = pdf_path
        self.raw_pdf_data = None
        self.pdf_reader = None
        
        try:
            with open(pdf_path, 'rb') as file:
                self.raw_pdf_data = file.read()
                self.pdf_reader = PyPDF2.PdfReader(io.BytesIO(self.raw_pdf_data))
                logger.info(f"Successfully loaded PDF: {os.path.basename(pdf_path)}")
        except Exception as e:
            logger.error(f"Error loading PDF: {e}")
            raise
    
    def extract_basic_info(self):
        """
        Extract basic information from the PDF
        
        Returns:
            dict: Basic information about the PDF
        """
        info = {
            'num_pages': len(self.pdf_reader.pages),
            'file_size': len(self.raw_pdf_data),
            'has_objects': len(self.pdf_reader.xref) > 0,
        }
        logger.info(f"Extracted basic info: {info}")
        return info
    
    def extract_streams(self):
        """
        Extract raw streams from the PDF which might contain embedded data
        
        Returns:
            list: List of stream data and their object IDs
        """
        streams = []
        
        # This is a simplified approach - in a real implementation,
        # you would need more sophisticated parsing of the PDF structure
        try:
            for i in range(len(self.pdf_reader.pages)):
                page = self.pdf_reader.pages[i]
                # Extract page contents which can contain streams
                if '/Contents' in page:
                    content = page['/Contents']
                    if isinstance(content, PyPDF2.generic.IndirectObject):
                        streams.append({
                            'page': i + 1,
                            'object_id': content.idnum,
                            'data_sample': binascii.hexlify(content.get_object().get_data()[:50])
                        })
            
            logger.info(f"Extracted {len(streams)} potential streams")
            return streams
        except Exception as e:
            logger.error(f"Error extracting streams: {e}")
            return []
            
    def extract_text_content(self):
        """
        Extract text content from the PDF
        
        Returns:
            str: Extracted text content
        """
        try:
            text = extract_text(self.pdf_path)
            logger.info(f"Extracted {len(text)} characters of text")
            return text
        except Exception as e:
            logger.error(f"Error extracting text: {e}")
            return ""
            
    def detect_potential_image_containers(self):
        """
        Detect objects in the PDF that might contain images or hidden data
        
        Returns:
            list: List of potential image containers with their properties
        """
        image_containers = []
        
        try:
            for page_layout in extract_pages(self.pdf_path):
                for element in page_layout:
                    if isinstance(element, (LTImage, LTFigure)):
                        container = {
                            'type': type(element).__name__,
                            'x0': element.x0,
                            'y0': element.y0,
                            'x1': element.x1,
                            'y1': element.y1,
                            'width': element.width,
                            'height': element.height,
                        }
                        image_containers.append(container)
            
            logger.info(f"Detected {len(image_containers)} potential image containers")
            return image_containers
        except Exception as e:
            logger.error(f"Error detecting image containers: {e}")
            return []
    
    def analyze_pdf_structure(self):
        """
        Analyze the overall PDF structure looking for anomalies
        
        Returns:
            dict: Analysis results
        """
        analysis = {
            'basic_info': self.extract_basic_info(),
            'streams': self.extract_streams(),
            'image_containers': self.detect_potential_image_containers(),
        }
        
        # Add summary statistics
        analysis['summary'] = {
            'num_streams': len(analysis['streams']),
            'num_image_containers': len(analysis['image_containers']),
        }
        
        return analysis