requirements.txt
asgiref
cffi
charset-normalizer
contourpy
cryptography
cycler
Deprecated
Django
fonttools
joblib
kiwisolver
lxml
matplotlib
numpy
packaging
pdfminer.six
pikepdf
pillow
pycparser
pyparsing
PyPDF2
python-dateutil
scikit-learn
scipy
six
sqlparse
threadpoolctl
tzdata==2025.2
wrapt
celery
redis
django-celery-results
djangorestframework
pandas
opencv-python
reportlab
pypdf2
pdfminer.six

torch
matplotlib
seaborn 
xgboost
numpy 
PyMuPDF 
opencv-python
tqdm
scikit-learn
tensorflow


